+++ chainer_chemistry/dataset/preprocessors/rsgcn_preprocessor.py
@@ -48,36 +48,20 @@ class RSGCNPreprocessor(MolPreprocessor):
 
         # adjust adjacent matrix
         degree_vec = numpy.sum(adj_array, axis=1)
-        degree_vec = numpy.sqrt(degree_vec)
-        adj_array *= numpy.broadcast_to(degree_vec[:, None], adj_array.shape)
-        adj_array *= numpy.broadcast_to(degree_vec[None, :], adj_array.shape)
+        degree_sqrt_inv = 1. / numpy.sqrt(degree_vec)
+        adj_array *= numpy.broadcast_to(degree_sqrt_inv[:, None],
+                                        adj_array.shape)
+        adj_array *= numpy.broadcast_to(degree_sqrt_inv[None, :],
+                                        adj_array.shape)
 
         return atom_array, adj_array
 
 
-class SparseRSGCNPreprocessor(MolPreprocessor):
+class SparseRSGCNPreprocessor(RSGCNPreprocessor):
     """Sparse RSGCN Preprocessor
 
-    Args:
-        max_atoms (int): Max number of atoms for each molecule, if the
-            number of atoms is more than this value, this data is simply
-            ignored.
-            Setting negative value indicates no limit for max atoms.
-        out_size (int): It specifies the size of array returned by
-            `get_input_features`.
-            If the number of atoms in the molecule is less than this value,
-            the returned arrays is padded to have fixed size.
-            Setting negative value indicates do not pad returned array.
-
+    See: RSGCN Preprocessor
     """
-
-    def __init__(self, max_atoms=-1, out_size=-1, add_Hs=False):
-        super(SparseRSGCNPreprocessor, self).__init__(add_Hs=add_Hs)
-        if max_atoms >= 0 and out_size >= 0 and max_atoms > out_size:
-            raise ValueError('max_atoms {} must be equal to or larger than '
-                             'out_size {}'.format(max_atoms, out_size))
-        self.max_atoms = max_atoms
-        self.out_size = out_size
 
     def get_input_features(self, mol):
         """get input features
@@ -88,17 +72,8 @@ class SparseRSGCNPreprocessor(MolPreprocessor):
         Returns:
 
         """
-        type_check_num_atoms(mol, self.max_atoms)
-        atom_array = construct_atomic_number_array(mol, out_size=self.out_size)
-        adj_array = construct_adj_matrix(mol, out_size=self.out_size)
-
-        # adjust adjacent matrix
-        degree_vec = numpy.sum(adj_array, axis=1)
-        degree_sqrt_inv = 1. / numpy.sqrt(degree_vec)
-        adj_array *= numpy.broadcast_to(degree_sqrt_inv[:, None],
-                                        adj_array.shape)
-        adj_array *= numpy.broadcast_to(degree_sqrt_inv[None, :],
-                                        adj_array.shape)
+        atom_array, adj_array = super(SparseRSGCNPreprocessor,
+                                      self).get_input_features(mol)
 
         # make sparse matrix
         sp_adj = sparse_dense2coo(adj_array)
