+++ chainer_chemistry/models/__init__.py
@@ -8,6 +8,6 @@ from chainer_chemistry.models.ggnn import GGNN  # NOQA
 from chainer_chemistry.models.mlp import MLP  # NOQA
 from chainer_chemistry.models.nfp import NFP  # NOQA
 from chainer_chemistry.models.rsgcn import RSGCN  # NOQA
-from chainer_chemistry.models.sparse_rsgcn import SparseRSGCN  # NOQA
+from chainer_chemistry.models.rsgcn import SparseRSGCN  # NOQA
 from chainer_chemistry.models.schnet import SchNet  # NOQA
 from chainer_chemistry.models.weavenet import WeaveNet  # NOQA
