--- chainer_chemistry/models/sparse_rsgcn.py
@@ -1,157 +1,0 @@
-"""
-Implementation of Renormalized Spectral Graph Convolutional Network (RSGCN)
-
-See: Thomas N. Kipf and Max Welling, \
-    Semi-Supervised Classification with Graph Convolutional Networks. \
-    September 2016. \
-    `arXiv:1609.02907 <https://arxiv.org/abs/1609.02907>`_
-"""
-import chainer
-from chainer import functions
-
-import chainer_chemistry
-from chainer_chemistry.config import MAX_ATOMIC_NUM
-
-
-class SparseRSGCNUpdate(chainer.Chain):
-    """Sparse RSGCN sub module for message and update part
-
-    Args:
-        in_channels (int): input channel dimension
-        out_channels (int): output channel dimension
-    """
-
-    def __init__(self, in_channels, out_channels):
-        super(SparseRSGCNUpdate, self).__init__()
-        with self.init_scope():
-            self.graph_linear = chainer_chemistry.links.GraphLinear(
-                in_channels, out_channels, nobias=True)
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-
-    def __call__(self, x, sp_adj):
-        # --- Message part ---
-        h = chainer_chemistry.functions.sparse_matmul(sp_adj, x)
-        # --- Update part ---
-        y = self.graph_linear(h)
-        return y
-
-
-def rsgcn_readout_sum(x, activation=None):
-    """Default readout function for `RSGCN`
-
-    Args:
-        x (chainer.Variable): shape consists of (minibatch, atom, ch).
-        activation: activation function, default is `None`.
-            You may consider taking other activations, for example `sigmoid`,
-            `relu` or `softmax` along `axis=2` (ch axis) etc.
-    Returns: result of readout, its shape should be (minibatch, out_ch)
-
-    """
-    if activation is not None:
-        h = activation(x)
-    else:
-        h = x
-    y = functions.sum(h, axis=1)  # sum along node axis
-    return y
-
-
-class SparseRSGCN(chainer.Chain):
-
-    """Renormalized Spectral Graph Convolutional Network (RSGCN)
-
-    See: Thomas N. Kipf and Max Welling, \
-        Semi-Supervised Classification with Graph Convolutional Networks. \
-        September 2016. \
-        `arXiv:1609.02907 <https://arxiv.org/abs/1609.02907>`_
-
-    The name of this model "Renormalized Spectral Graph Convolutional Network
-    (RSGCN)" is named by us rather than the authors of the paper above.
-    The authors call this model just "Graph Convolution Network (GCN)", but
-    we think that "GCN" is bit too general and may cause namespace issue.
-    That is why we did not name this model as GCN.
-
-    Args:
-        out_dim (int): dimension of output feature vector
-        hidden_dim (int): dimension of feature vector
-            associated to each atom
-        n_atom_types (int): number of types of atoms
-        n_layers (int): number of layers
-        use_batch_norm (bool): If True, batch normalization is applied after
-            graph convolution.
-        readout (Callable): readout function. If None, `rsgcn_readout_sum` is
-            used. To the best of our knowledge, the paper of RSGCN model does
-            not give any suggestion on readout.
-    """
-
-    def __init__(self, out_dim, hidden_dim=32, n_layers=4,
-                 n_atom_types=MAX_ATOMIC_NUM,
-                 use_batch_norm=False, readout=None):
-        super(SparseRSGCN, self).__init__()
-        in_dims = [hidden_dim for _ in range(n_layers)]
-        out_dims = [hidden_dim for _ in range(n_layers)]
-        out_dims[n_layers - 1] = out_dim
-        self.readout = None
-        with self.init_scope():
-            self.embed = chainer_chemistry.links.EmbedAtomID(
-                in_size=n_atom_types, out_size=hidden_dim)
-            self.gconvs = chainer.ChainList(
-                *[SparseRSGCNUpdate(in_dims[i], out_dims[i])
-                  for i in range(n_layers)])
-            if use_batch_norm:
-                self.bnorms = chainer.ChainList(
-                    *[chainer_chemistry.links.GraphBatchNormalization(
-                        out_dims[i]) for i in range(n_layers)])
-            else:
-                self.bnorms = [None for _ in range(n_layers)]
-            if isinstance(readout, chainer.Link):
-                self.readout = readout
-        if self.readout is None:
-            self.readout = readout or rsgcn_readout_sum
-        self.out_dim = out_dim
-        self.hidden_dim = hidden_dim
-        self.n_layers = n_layers
-
-    def __call__(self, graph, adj_data, adj_row, adj_col):
-        """Forward propagation
-
-        Args:
-            graph (numpy.ndarray): minibatch of molecular which is
-                represented with atom IDs (representing C, O, S, ...)
-                `atom_array[mol_index, atom_index]` represents `mol_index`-th
-                molecule's `atom_index`-th atomic number
-            adj_data (numpy.ndarray): minibatch of adjacency matrix.
-            adj_row (numpy.ndarray): minibatch of adjacency matrix.
-            adj_col (numpy.ndarray): minibatch of adjacency matrix.
-                COO format is adopted as sparse matrix format. adj_data,
-                adj_row and adj_col are data, row index and column index array
-                of the matrix respectively.
-
-        Returns:
-            ~chainer.Variable: minibatch of fingerprint
-        """
-        if graph.dtype == self.xp.int32:
-            # atom_array: (minibatch, nodes)
-            h = self.embed(graph)
-        else:
-            h = graph
-        # h: (minibatch, nodes, ch)
-
-        adj_shape = [h.shape[1], h.shape[1]]
-        sp_adj = chainer_chemistry.functions.sparse_coo_matrix(
-            adj_data, adj_row, adj_col, adj_shape)
-        # adj_data/row/col: (minibatch, nnz)
-
-        # --- Sparse RSGCN update ---
-        for i, (gconv, bnorm) in enumerate(zip(self.gconvs,
-                                               self.bnorms)):
-            h = gconv(h, sp_adj)
-            if bnorm is not None:
-                h = bnorm(h)
-            h = functions.dropout(h)
-            if i < self.n_layers - 1:
-                h = functions.relu(h)
-
-        # --- readout ---
-        y = self.readout(h)
-        return y
